move `parseLine` to utils
